[respeaker_ros] importing does not succeed
